+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Yii2 Queue Extension Change Log
 2.3.1 December 23, 2020
 -----------------------
 
+- Enh #372: Add ability to configure keepalive and heartbeat for Amqp (vyachin)
 - Bug #380: Fixed amqp-interop queue/listen signal handling (tarinu)
 - Enh #388: `symfony/process 5.0` compatibility (leandrogehlen)